// masjid-core/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  SUSPENDED: 'SUSPENDED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const RoleName = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  MASJID_ADMIN: 'MASJID_ADMIN',
  IMAM: 'IMAM',
  COMMITTEE_MEMBER: 'COMMITTEE_MEMBER',
  MEMBER: 'MEMBER'
} as const

export type RoleName = (typeof RoleName)[keyof typeof RoleName]


export const PaymentStatus = {
  PAID: 'PAID',
  UNPAID: 'UNPAID',
  PARTIAL: 'PARTIAL'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const ImamSalaryAssignmentStatus = {
  UNPAID: 'UNPAID',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID'
} as const

export type ImamSalaryAssignmentStatus = (typeof ImamSalaryAssignmentStatus)[keyof typeof ImamSalaryAssignmentStatus]


export const PaymentMode = {
  CASH: 'CASH',
  ONLINE: 'ONLINE'
} as const

export type PaymentMode = (typeof PaymentMode)[keyof typeof PaymentMode]


export const ProjectStatus = {
  PLANNED: 'PLANNED',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const CollectionType = {
  JUMMA_COLLECTION: 'JUMMA_COLLECTION',
  DONATION_BOX: 'DONATION_BOX',
  RAMADAN_FUND: 'RAMADAN_FUND',
  ZAKAT: 'ZAKAT',
  SADAQAH: 'SADAQAH',
  CONSTRUCTION_FUND: 'CONSTRUCTION_FUND',
  OTHER: 'OTHER'
} as const

export type CollectionType = (typeof CollectionType)[keyof typeof CollectionType]


export const ExpenseType = {
  ELECTRICITY_BILL: 'ELECTRICITY_BILL',
  WATER_BILL: 'WATER_BILL',
  IMAM_SALARY: 'IMAM_SALARY',
  CLEANING: 'CLEANING',
  REPAIR: 'REPAIR',
  CONSTRUCTION: 'CONSTRUCTION',
  OTHER: 'OTHER'
} as const

export type ExpenseType = (typeof ExpenseType)[keyof typeof ExpenseType]


export const FinanceEntryStatus = {
  ACTIVE: 'ACTIVE',
  CANCELLED: 'CANCELLED'
} as const

export type FinanceEntryStatus = (typeof FinanceEntryStatus)[keyof typeof FinanceEntryStatus]


export const MasjidStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  SUSPENDED: 'SUSPENDED'
} as const

export type MasjidStatus = (typeof MasjidStatus)[keyof typeof MasjidStatus]


export const MasjidRegistrationStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type MasjidRegistrationStatus = (typeof MasjidRegistrationStatus)[keyof typeof MasjidRegistrationStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]